Separate GetCompileErrors from ClearConsoleLogs and GetCurrentConsoleLogs
