+++ UnityNaturalMCPServer/Editor/McpTools/McpUnityEditorTool.cs
@@ -1,23 +1,19 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.ComponentModel;
-using System.Linq;
-using System.Reflection;
-using System.Text.RegularExpressions;
 using System.Threading.Tasks;
 using Cysharp.Threading.Tasks;
 using ModelContextProtocol.Server;
 using UnityEditor;
 using UnityEditor.TestTools.TestRunner.Api;
 using UnityEngine;
-using Assert = UnityEngine.Assertions.Assert;
 
 namespace UnityNaturalMCP.Editor.McpTools
 {
     [McpServerToolType, Description("Control Unity Editor tools")]
     internal sealed class McpUnityEditorTool
     {
-        [McpServerTool, Description("Execute AssetDatabase.Refresh")]
+        [McpServerTool, Description("Execute AssetDatabase.Refresh.It must also be called when compiling a script.")]
         public async ValueTask RefreshAssets()
         {
             try
@@ -32,7 +28,7 @@ namespace UnityNaturalMCP.Editor.McpTools
             }
         }
 
-        [McpServerTool, Description("Get current console logs. Recommend calling ClearConsoleLogs beforehand.")]
+        [McpServerTool, Description("Get current console logs.The default value is adjusted to reduce token usage. Simply call it with the default value the first time and specify arguments only when necessary.")]
         public async ValueTask<IReadOnlyList<LogEntry>> GetCurrentConsoleLogs(
             [Description(
                 "Filter logs by type. Valid values: \"\"(Maches all logs), \"error\", \"warning\", \"log\", \"compile-error\"(This is all you need to check for compilation errors.), \"compile-warning\"")]
@@ -41,7 +37,8 @@ namespace UnityNaturalMCP.Editor.McpTools
             string filter = "",
             [Description("Log count limit. Set to 0 for no limit(Not recommended).")]
             int maxCount = 20,
-            [Description("Get only first line of the log message. If false, the whole message is returned.(To save tokens, recommend calling this with true.)")]
+            [Description(
+                "Get only first line of the log message. If false, the whole message is returned.")]
             bool onlyFirstLine = true,
             [Description(
                 "If true, the logs will be sorted by time in chronological order(oldest first). If false, newest first.")]
@@ -50,51 +47,8 @@ namespace UnityNaturalMCP.Editor.McpTools
             try
             {
                 await UniTask.SwitchToMainThread();
-                var logTypeToLower = logType.ToLower();
-                var logs = new List<LogEntry>();
-                var logEntries = Type.GetType("UnityEditor.LogEntries,UnityEditor.dll");
-                Assert.IsNotNull(logEntries);
 
-                var getCountMethod = logEntries.GetMethod("GetCount", BindingFlags.Public | BindingFlags.Static);
-                var getEntryInternalMethod = logEntries.GetMethod("GetEntryInternal", BindingFlags.Public | BindingFlags.Static);
-
-                Assert.IsNotNull(getCountMethod);
-                Assert.IsNotNull(getEntryInternalMethod);
-
-                var count = (int)getCountMethod.Invoke(null, null);
-
-                for (var i = 0; i < count; i++)
-                {
-                    var logEntryType = Type.GetType("UnityEditor.LogEntry,UnityEditor.dll");
-                    Assert.IsNotNull(logEntryType);
-
-                    var logEntry = Activator.CreateInstance(logEntryType);
-
-                    getEntryInternalMethod.Invoke(null, new[] { i, logEntry });
-
-                    var message = logEntry.GetType().GetField("message").GetValue(logEntry) as string ?? "";
-                    var file = logEntry.GetType().GetField("file").GetValue(logEntry) as string ?? "";
-                    var mode = (int)logEntry.GetType().GetField("mode").GetValue(logEntry);
-                    var logTypeValue = UnityInternalLogModeToLogType(mode);
-
-                    if ((string.IsNullOrEmpty(logTypeToLower) || logTypeValue.Equals(logTypeToLower))
-                        && (string.IsNullOrEmpty(filter) || Regex.IsMatch(message, filter)))
-                    {
-                        logs.Add(new LogEntry(onlyFirstLine ? message.Split('\n')[0] : message, logTypeValue));
-                    }
-                }
-
-                if (!isChronological)
-                {
-                    logs = ((IEnumerable<LogEntry>)logs).Reverse().ToList();
-                }
-
-                if (maxCount > 0)
-                {
-                    logs = logs.Take(maxCount).ToList();
-                }
-
-                return logs;
+                return ConsoleLogUtilities.GetLogs(filter, maxCount, onlyFirstLine, isChronological, logType.ToLower());
             }
             catch (Exception e)
             {
@@ -103,20 +57,42 @@ namespace UnityNaturalMCP.Editor.McpTools
             }
         }
 
-        [McpServerTool, Description("Clear console logs. It is recommended to call it before GetCurrentConsoleLogs.")]
+        [McpServerTool, Description("Clear console logs.")]
         public async ValueTask ClearConsoleLogs()
         {
             try
             {
                 await UniTask.SwitchToMainThread();
-                var logEntries = Type.GetType("UnityEditor.LogEntries,UnityEditor.dll");
-                Assert.IsNotNull(logEntries);
+                ConsoleLogUtilities.ClearLogs();
+            }
+            catch (Exception e)
+            {
+                Debug.LogError(e);
+                throw;
+            }
+        }
 
-                var clearMethod = logEntries.GetMethod("Clear", BindingFlags.Public | BindingFlags.Static);
+        [McpServerTool,
+         Description(
+             "Get compilation errors. Same as `ClearConsoleLgs();GetCurrentConsoleLogs(\"compile-error\", args)`")]
+        public async ValueTask<IReadOnlyList<LogEntry>> GetCompileErrors(
+            [Description("Filter by regex. If empty, all logs are returned.")]
+            string filter = "",
+            [Description("Log count limit. Set to 0 for no limit(Not recommended).")]
+            int maxCount = 20,
+            [Description(
+                "Get only first line of the log message. If false, the whole message is returned.(To save tokens, recommend calling this with true.)")]
+            bool onlyFirstLine = true,
+            [Description(
+                "If true, the logs will be sorted by time in chronological order(oldest first). If false, newest first.")]
+            bool isChronological = false)
+        {
+            try
+            {
+                await UniTask.SwitchToMainThread();
 
-                Assert.IsNotNull(clearMethod);
-
-                clearMethod.Invoke(null, null);
+                ConsoleLogUtilities.ClearLogs();
+                return ConsoleLogUtilities.GetLogs(filter, maxCount, onlyFirstLine, isChronological, "compile-error");
             }
             catch (Exception e)
             {
@@ -165,15 +141,5 @@ namespace UnityNaturalMCP.Editor.McpTools
                 }
             }
         }
-
-        private string UnityInternalLogModeToLogType(int mode) => mode switch
-        {
-            _ when (mode & (int)LogMessageFlags.ScriptingError) != 0 => "error",
-            _ when (mode & (int)LogMessageFlags.ScriptingWarning) != 0 => "warning",
-            _ when (mode & (int)LogMessageFlags.ScriptingLog) != 0 => "log",
-            _ when (mode & (int)LogMessageFlags.ScriptCompileError) != 0 => "compile-error",
-            _ when (mode & (int)LogMessageFlags.ScriptCompileWarning) != 0 => "compile-warning",
-            _ => "unknown"
-        };
     }
 }
